+++ src/actions/index.js
@@ -1,88 +1,8 @@
 import * as types from  '../constants/ActionTypes'
-import axios from 'axios';
-import loadGoogleMapsAPI from 'load-google-maps-api';
-
-const mapsapi = loadGoogleMapsAPI({
-    key: process.env.REACT_APP_GOOGLE_KEY,
-    libraries: ['places'],
-    language: 'fr'
-});
-
-const receiveAutocompleteResponse = (autocomplete_name, items, request_time) => {
-    return {
-        type: types.RECEIVE_AUTOCOMPLETE_RESPONSE,
-        items,
-        autocomplete: autocomplete_name,
-        request_time
-    }
-}
-
-const receiveAutocompleteError= (autocomplete_name, error_message) => {
-    return {
-        type: types.RECEIVE_AUTOCOMPLETE_ERROR,
-        error: error_message,
-        autocomplete: autocomplete_name
-    }
-}
-
-const sendRequest = (url, params, autocomplete, 
-    successCallback, errorCallback = null, headers = {}) => dispatch => {
-    const startTime = new Date().getTime();
-    return axios.get(url, {params, headers})
-        .then(response => {
-            dispatch(receiveAutocompleteResponse(
-                autocomplete,
-                successCallback(response),
-                new Date().getTime() - startTime
-            ));
-        })
-        .catch(error => {
-            if (error.response && typeof errorCallback === 'function') {
-                dispatch(receiveAutocompleteError(
-                    autocomplete,
-                    errorCallback(error.response.data)
-                ));
-            } else {
-                dispatch(receiveAutocompleteError(autocomplete, error.message));
-            }
-        });
-};
-
-const sendRequestGooglePlaces = (term, autocomplete, successCallback) => dispatch => {
-    mapsapi
-        .then(googleMaps => {
-            return new window.google.maps.places.AutocompleteService();
-        }).then(autocompleteService => {
-            const startTime = new Date().getTime();
-            autocompleteService.getPlacePredictions(
-                { input: term},
-                (predictions, status) => {
-                    if (status === window.google.maps.places.PlacesServiceStatus.OK) {
-                        dispatch(receiveAutocompleteResponse(
-                            'google',
-                            predictions.map(prediction => {
-                                return {
-                                    label: prediction.description,
-                                    type: null
-                                };
-                            }),
-                            new Date().getTime() - startTime
-                        ));
-                    } else if (status === window.google.maps.places.PlacesServiceStatus.ZERO_RESULTS) {
-                        dispatch(receiveAutocompleteResponse(
-                            'google',
-                            [],
-                            0
-                        ));
-                    } else {
-                        dispatch(receiveAutocompleteError('google', status));
-                    }
-                }
-            )
-        }).catch((err) => {
-            console.error(err);
-        });
-};
+import requestBragi from './bragi';
+import requestKraken from './kraken';
+import requestBano from './bano';
+import requestGooglePlaces from './google';
 
 export const handleInputChange = (autocomplete_name, name, value) => {
     return {
@@ -93,64 +13,23 @@ export const handleInputChange = (autocomplete_name, name, value) => {
     }
 }
 
+export const changeSearchTerm = (term) => {
+    return {
+        type: types.CHANGE_SEARCH_TERM,
+        term
+    }
+}
+
 export const requestAutocompletes = (term = null) => {
     const thunk = (dispatch, getState) => {
         if (getState().term === '') {
             return;
         }
         Promise.all([
-            dispatch(sendRequest(
-                `${getState().bragi.inputs.bragi_host}/autocomplete`,
-                { q: term},
-                'bragi',
-                response => response.data.features.map(feature => {
-                    let type;
-                    const geocoding = feature.properties.geocoding;
-                    if (geocoding.type === 'public_transport:stop_area') {
-                        type = 'stop_area';
-                    } else if (geocoding.type === 'house' || geocoding.type === 'street') {
-                        type = 'address';
-                    } else {
-                        type = geocoding.type;
-                    }
-
-                    return {
-                        label: feature.properties.geocoding.label,
-                        type
-                    };
-                })
-            )),
-            dispatch(sendRequest(
-                `${getState().kraken.inputs.navitia_host}/v1/coverage/${getState().kraken.inputs.navitia_coverage}/places`,
-                { q: term},
-                'kraken',
-                response => {
-                    if (response.data.hasOwnProperty("places")) {
-                        return response.data.places.map(place => {
-                            return {
-                                label: place.name,
-                                type: null
-                            };
-                        });
-                    }
-
-                    return [];
-                },
-                err => err.message,
-                { 'Authorization': getState().kraken.inputs.navitia_token }
-            )),
-            dispatch(sendRequest(
-                'http://api-adresse.data.gouv.fr/search',
-                { q: term },
-                'bano',
-                response => response.data.features.map(feature => {
-                    return {
-                        label: feature.properties.label,
-                        type: null
-                    };
-                })
-            )),
-            dispatch(sendRequestGooglePlaces(term, 'google')),
+            dispatch(requestBragi(term)),
+            dispatch(requestKraken(term)),
+            dispatch(requestBano(term)),
+            dispatch(requestGooglePlaces(term)),
         ]);
     }
 
@@ -162,11 +41,4 @@ export const requestAutocompletes = (term = null) => {
     };
 
     return thunk;
-}
-
-export const changeSearchTerm = (term) => {
-    return {
-        type: types.CHANGE_SEARCH_TERM,
-        term
-    }
 }
